+++ src/components/Order/Order.css
@@ -2,15 +2,15 @@
   display: flex;
 
   flex-direction: column;
-  margin: 20px 0px 100px 40px;
+  margin: 20px 40px 100px 40px;
 
   justify-content: space-between;
 }
 
 .order__content {
-  display: grid;
-  grid-template-columns: auto auto;
-  flex-direction: column;
+  display: flex;
+  flex-direction: row;
+
   justify-content: space-between;
   align-items: baseline;
 }
@@ -22,9 +22,9 @@
 }
 
 .order__link {
-  display: grid;
-  margin-top: 60px;
+  margin: 0 0 18px 0;
   cursor: pointer;
+  max-width: 200px;
 }
 
 .order__link-text {
@@ -53,31 +53,7 @@
   line-height: 56px;
   /* 116.667% */
   letter-spacing: -1px;
-  margin: 0 0 -60px 0;
-}
-
-.order__button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  align-self: stretch;
-  border-radius: 6px;
-  background: var(--black-900, #121212);
-  color: var(--White, #fff);
-  text-align: center;
-  font-feature-settings:
-    'clig' off,
-    'liga' off;
-  /* Button/Large */
-  font-family: Space Grotesk;
-  font-size: 22px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 34px;
-  /* 154.545% */
-  width: 434px;
-  height: 72px;
-  cursor: pointer;
+  margin: 40px 0 32px 0;
 }
 
 .order__button-link {
@@ -86,25 +62,4 @@
 
 .order__arrow {
   margin: 0 16px 0 0;
-}
-
-.order__politic {
-  color: #000;
-  font-family: Space Grotesk;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-  /* 166.667% */
-}
-
-.order__politic-text {
-  text-decoration: none;
-  color: #000;
-  font-family: Space Grotesk;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-  /* 166.667% */
 }
