+++ src/components/OrderProduct/OrderProduct.css
@@ -1,21 +1,21 @@
 .orderproduct {
   list-style-type: none;
-  margin: 24px 0 0 24px;
+  margin: 24px 24px 0 24px;
   padding: 0;
-  border-bottom: 1px solid var(--black-300, #cbcbcb);
+  /* max-width: 548px; */
 }
 
 .orderproduct__product {
   display: grid;
-  grid-template-columns: 84px 160px auto auto auto;
+  grid-template-columns: 84px minmax(auto, 276px) auto auto;
   grid-template-rows: auto auto 16px;
-  width: 384px;
+  /* max-width: 500px; */
   height: 100px;
 }
 
 .orderproduct__image {
   width: 84px;
-  height: 100px;
+  height: 84px;
   grid-column-start: 1;
   grid-row-start: 1;
   grid-row-end: 3;
@@ -25,7 +25,7 @@
   grid-column-start: 2;
   grid-row-start: 1;
   margin: 0 0 0 16px;
-  width: 160px;
+  /* width: 160px; */
   height: 24px;
   align-self: end;
   font-family: Space Grotesk;
@@ -39,7 +39,7 @@
   grid-column-start: 2;
   grid-row-start: 2;
   margin: 0 0 0 16px;
-  width: 160px;
+  /* width: 160px; */
   height: 24px;
   color: var(--black-600, #605f5f);
   font-feature-settings:
@@ -59,7 +59,7 @@
   grid-row-start: 1;
   grid-row-end: 3;
   align-self: center;
-  width: 38px;
+  /* width: 38px; */
   color: var(--black-900, #121212);
   text-align: right;
   font-feature-settings:
@@ -74,7 +74,7 @@
   /* margin-right: 16px; */
 }
 
-.orderproduct__price {
+/* .orderproduct__price {
   grid-column-start: 4;
   grid-row-start: 1;
   grid-row-end: 3;
@@ -86,19 +86,19 @@
     'clig' off,
     'liga' off;
   /* Text/14px/Semibold */
-  font-family: Inter;
+/* font-family: Inter;
   font-size: 14px;
   font-style: normal;
   font-weight: 600;
-  line-height: 22px; /* 157.143% */
-}
+  line-height: 22px;
+} */
 
 .orderproduct__totalprice {
-  grid-column-start: 5;
+  grid-column-start: 4;
   grid-row-start: 1;
   grid-row-end: 3;
   align-self: center;
-  width: 58px;
+  /* width: 58px; */
   color: var(--black-900, #121212);
   text-align: right;
   font-feature-settings:
