+++ src/components/OrderProduct/OrderProduct.jsx
@@ -2,12 +2,18 @@ import './OrderProduct.css';
 import { useState } from 'react';
 
 function OrderProduct({ product }) {
+  const loc = useState(
+    JSON.parse(localStorage.getItem(`purchaseItem${product.id}`))
+  );
   const counter = useState(
-    JSON.parse(localStorage.getItem(`purchaseItem${product.id}`)).counter
+    JSON.parse(localStorage.getItem(`purchaseItem${product.id}`)).amount
   );
   const totalItemPrice = useState(
     JSON.parse(localStorage.getItem(`purchaseItem${product.id}`)).totalItemPrice
   );
+  console.log(counter);
+  console.log(totalItemPrice);
+  console.log(loc);
 
   //  const totalItemPrice =  useState(product.price * counter)
 
@@ -22,11 +28,9 @@ function OrderProduct({ product }) {
         <p className="odredproduct__count">{counter} шт</p>
         <p className="orderproduct__name">{product.brand}</p>
         <p className="orderproduct__description">{'Детали(объем, вес)'}</p>
-        <p className="orderproduct__price">{product.price} ₽</p>
+        {/* <p className="orderproduct__price">{product.price} ₽</p> */}
         <p className="orderproduct__totalprice">{totalItemPrice} ₽</p>
       </li>
-
-      <br className="orderproduct__line"></br>
     </ul>
   );
 }
