+++ src/components/YourOrder/YourOrder.css
@@ -1,12 +1,11 @@
 .yourorder {
   display: flex;
-  max-width: 432px;
+  /* max-width: 548px; */
   flex-direction: column;
-  align-items: flex-start;
+  align-items: stretch;
   border-radius: 6px;
   border: 1px solid var(--black-300, #cbcbcb);
   background: var(--White, #fff);
-  margin: -18px 0 40px 0;
 }
 
 .yourorder__title {
@@ -25,14 +24,14 @@
 
 .yourorder__list {
   padding: 0;
+  /* max-width: 500px; */
 }
 
 .yourorder__total {
   display: flex;
   margin: 28px 24px 16px 24px;
-  width: 384px;
+  max-width: 500px;
   justify-content: space-between;
-  height: 54px;
 }
 
 .yourorder__total-text {
@@ -56,5 +55,54 @@
   font-style: normal;
   font-weight: 600;
   line-height: 30px;
+  /* 166.667% */
+}
+
+.yourorder__button {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  align-self: stretch;
+  border-radius: 12px;
+  background: var(--bright-logo-color, #a49e4a);
+  color: var(--White, #fff);
+  text-align: center;
+  font-feature-settings:
+    'clig' off,
+    'liga' off;
+  /* Button/Large */
+  font-family: Space Grotesk;
+  font-size: 22px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 34px;
+  /* 154.545% */
+  /* width: 500px; */
+  height: 56px;
+  cursor: pointer;
+  margin: 0 24px 0 24px;
+}
+
+.yourorder__politic {
+  margin: 0px 24px 0px 24px;
+  color: #000;
+  font-family: Space Grotesk;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 20px;
+  align-self: center;
+  /* 166.667% */
+}
+
+.yourorder__politic-text {
+  text-decoration: none;
+  color: #000;
+  font-family: Space Grotesk;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 20px;
+
   /* 166.667% */
 }
